Removing Php from Code, Added Initial Graphical Veiw
In order to make the pages static the process of Php removal from is
started, also the Initial Graphical Veiw has been updated.

Features added are :-

1. Options to add/edit nodes and edges
2. Directed Edges
3. Project Level Dependencies

A node with "unknown" label tells that the project to which the actual
dependency belonged to could not be identified. In later versions this
issue will be resolved.